Meta Tags & Error Checking
So, the meta tags are now added to make it a bit more "professional"

The error checking was just the CPanel JSON response if it does not come back with anything just redirect to services.php. This will do for now and perhaps in the future I'll add a correct error check.